+++ lib/gen/l10n/app_localizations.dart
@@ -98,23 +98,17 @@ abstract class AppLocalizations {
     Locale('fr'),
   ];
 
-  /// No description provided for @appTitle.
+  /// DO NOT TRANSLATE. This is the official brand name.
   ///
   /// In en, this message translates to:
   /// **'VTeam Cards'**
   String get appTitle;
 
-  /// No description provided for @cardsTitle.
+  /// DO NOT TRANSLATE.
   ///
   /// In en, this message translates to:
   /// **'Cards'**
   String get cardsTitle;
-
-  /// No description provided for @welcomeTitle.
-  ///
-  /// In en, this message translates to:
-  /// **'VTeam Cards'**
-  String get welcomeTitle;
 
   /// No description provided for @startNewGame.
   ///
@@ -488,13 +482,13 @@ abstract class AppLocalizations {
   /// **'9 Cards'**
   String get golf9Cards;
 
-  /// No description provided for @skyJo.
+  /// DO NOT TRANSLATE. This is the official brand name.
   ///
   /// In en, this message translates to:
-  /// **'SkyJo'**
-  String get skyJo;
+  /// **'Skyjo'**
+  String get skyjo;
 
-  /// No description provided for @miniPut.
+  /// DO NOT TRANSLATE.
   ///
   /// In en, this message translates to:
   /// **'MiniPut'**
@@ -511,12 +505,6 @@ abstract class AppLocalizations {
   /// In en, this message translates to:
   /// **'MiniPut 4 Cards'**
   String get miniPutFull;
-
-  /// No description provided for @skyLo.
-  ///
-  /// In en, this message translates to:
-  /// **'Skyjo'**
-  String get skyLo;
 
   /// No description provided for @columnsByRows.
   ///
@@ -668,12 +656,6 @@ abstract class AppLocalizations {
   /// **'Oh NO!'**
   String get statusOhNo;
 
-  /// No description provided for @playerNumber.
-  ///
-  /// In en, this message translates to:
-  /// **'Player {number}'**
-  String playerNumber(int number);
-
   /// No description provided for @nameForPlayerNumber.
   ///
   /// In en, this message translates to:
@@ -704,11 +686,11 @@ abstract class AppLocalizations {
   /// **'- Aim for the lowest score.\n- Choose a card from either the Deck or Discard pile.\n- Swap the chosen card with a card in your 3x3 grid, or discard it and flip over one of your face-down cards.\n- Three cards of the same rank in a row or column score zero.\n- The first player to reveal all nine cards challenges others, claiming the lowest score.\n- If someone else has an equal or lower score, the challenger doubles their points!\n- Players are eliminated after busting 100 points.\n\n\nLearn more [Wikipedia](https://en.wikipedia.org/wiki/Golf_(card_game))'**
   String get instructionsFrenchCards9;
 
-  /// No description provided for @instructionsSkyJo.
+  /// No description provided for @instructionsSkyjo.
   ///
   /// In en, this message translates to:
-  /// **'- Aim for the lowest score.\n- Choose a card from either the Deck or Discard pile.\n- Swap the chosen card with a card in your 4x3 grid, or discard it and flip over one of your face-down cards.\n- When 3 cards of the same rank are lined up in a column they are moved to the discard pile.\n- The first player to reveal all their cards challenges others, claiming the lowest score.\n\n\nLearn more [SkyJo](https://www.geekyhobbies.com/how-to-play-skyjo-card-game-rules-and-instructions/)'**
-  String get instructionsSkyJo;
+  /// **'- Aim for the lowest score.\n- Choose a card from either the Deck or Discard pile.\n- Swap the chosen card with a card in your 4x3 grid, or discard it and flip over one of your face-down cards.\n- When 3 cards of the same rank are lined up in a column they are moved to the discard pile.\n- The first player to reveal all their cards challenges others, claiming the lowest score.\n\n\nLearn more [Skyjo](https://www.geekyhobbies.com/how-to-play-skyjo-card-game-rules-and-instructions/)'**
+  String get instructionsSkyjo;
 
   /// No description provided for @instructionsMiniPut.
   ///