+++ lib/gen/l10n/app_localizations_en.dart
@@ -15,9 +15,6 @@ class AppLocalizationsEn extends AppLocalizations {
   String get cardsTitle => 'Cards';
 
   @override
-  String get welcomeTitle => 'VTeam Cards';
-
-  @override
   String get startNewGame => 'Start a New Game';
 
   @override
@@ -225,7 +222,7 @@ class AppLocalizationsEn extends AppLocalizations {
   String get golf9Cards => '9 Cards';
 
   @override
-  String get skyJo => 'SkyJo';
+  String get skyjo => 'Skyjo';
 
   @override
   String get miniPut => 'MiniPut';
@@ -235,9 +232,6 @@ class AppLocalizationsEn extends AppLocalizations {
 
   @override
   String get miniPutFull => 'MiniPut 4 Cards';
-
-  @override
-  String get skyLo => 'Skyjo';
 
   @override
   String columnsByRows(int columns, int rows) {
@@ -325,11 +319,6 @@ class AppLocalizationsEn extends AppLocalizations {
   String get statusOhNo => 'Oh NO!';
 
   @override
-  String playerNumber(int number) {
-    return 'Player $number';
-  }
-
-  @override
   String nameForPlayerNumber(int number) {
     return 'Name for Player #$number';
   }
@@ -350,8 +339,8 @@ class AppLocalizationsEn extends AppLocalizations {
       '- Aim for the lowest score.\n- Choose a card from either the Deck or Discard pile.\n- Swap the chosen card with a card in your 3x3 grid, or discard it and flip over one of your face-down cards.\n- Three cards of the same rank in a row or column score zero.\n- The first player to reveal all nine cards challenges others, claiming the lowest score.\n- If someone else has an equal or lower score, the challenger doubles their points!\n- Players are eliminated after busting 100 points.\n\n\nLearn more [Wikipedia](https://en.wikipedia.org/wiki/Golf_(card_game))';
 
   @override
-  String get instructionsSkyJo =>
-      '- Aim for the lowest score.\n- Choose a card from either the Deck or Discard pile.\n- Swap the chosen card with a card in your 4x3 grid, or discard it and flip over one of your face-down cards.\n- When 3 cards of the same rank are lined up in a column they are moved to the discard pile.\n- The first player to reveal all their cards challenges others, claiming the lowest score.\n\n\nLearn more [SkyJo](https://www.geekyhobbies.com/how-to-play-skyjo-card-game-rules-and-instructions/)';
+  String get instructionsSkyjo =>
+      '- Aim for the lowest score.\n- Choose a card from either the Deck or Discard pile.\n- Swap the chosen card with a card in your 4x3 grid, or discard it and flip over one of your face-down cards.\n- When 3 cards of the same rank are lined up in a column they are moved to the discard pile.\n- The first player to reveal all their cards challenges others, claiming the lowest score.\n\n\nLearn more [Skyjo](https://www.geekyhobbies.com/how-to-play-skyjo-card-game-rules-and-instructions/)';
 
   @override
   String get instructionsMiniPut =>